requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy[asyncio]==2.0.38
asyncpg==0.30.0
alembic==1.14.1
passlib[bcrypt]==1.7.4
bcrypt<4.0
python-jose[cryptography]==3.3.0
